post를 부분부분으로 가져오기
