+++ src/util/api/notion/block.ts
@@ -1,11 +1,13 @@
 import { ListBlockChildrenResponse } from "@notionhq/client/build/src/api-endpoints";
 
 export const RetrieveBlockChildren = async (id: string, start_cursor?: string): Promise<ListBlockChildrenResponse> => {
-    let url = `https://api.notion.com/v1/blocks/${id}/children?page_size=100`;
+    let url = `https://api.notion.com/v1/blocks/${id}/children?page_size=25`;
 
     if (start_cursor) {
-        url = `https://api.notion.com/v1/blocks/${id}/children?page_size=100&start_cursor=${start_cursor}`;
+        url = `https://api.notion.com/v1/blocks/${id}/children?page_size=25&start_cursor=${start_cursor}`;
     }
+
+    console.log(url);
 
     const res = await (
         await fetch(url, {
